restructiure CommentReaction model and migration
