+++ controllers/index.js
@@ -5,3 +5,4 @@ export { default as Profile } from './profile';
 export { default as Follow } from './follow';
 export { default as Comment } from './comment';
 export { default as Reaction } from './reaction';
+export { default as CommentReaction } from './commentReaction';
