+++ controllers/reaction.js
@@ -2,7 +2,7 @@ import Sequelize from 'sequelize';
 import models from '../models';
 
 const { iLike } = Sequelize.Op;
-const { Article, Reaction, commentReaction } = models;
+const { Article, Reaction } = models;
 
 const setReaction = async (liked, disliked, slug, userId) => {
   const article = await Article.findOne({
@@ -26,26 +26,6 @@ const setReaction = async (liked, disliked, slug, userId) => {
   });
 };
 
-const setCommentReaction = async (liked, disliked, commentId, userId) => {
-  const oldCommentReaction = await commentReaction.findOne({
-    where: {
-      commentId,
-      commentLikedBy: userId,
-    }
-  });
-  if (oldCommentReaction) {
-    await oldCommentReaction.update({ liked: disliked });
-    return;
-  }
-  await commentReaction.create({
-    commentId,
-    commentLikedBy: userId,
-    liked,
-    disliked
-  });
-};
-
-
 export default {
   async like(req, res) {
     await setReaction(true, false, req.params.slug, req.decoded.id);
@@ -60,13 +40,5 @@ export default {
   async dislike(req, res) {
     await setReaction(false, true, req.params.slug, req.decoded.id);
     return res.sendStatus(200);
-  },
-  async likeComment(req, res) {
-    await setCommentReaction(true, false, req.params.commentId, req.decoded.id);
-    return res.sendStatus(200);
-  },
-  async dislikeComment(req, res) {
-    await setCommentReaction(false, true, req.params.commentId, req.decoded.id);
-    return res.sendStatus(200);
-  },
+  }
 };
